# docs/swagger.yaml
definitions:
...
  dto.CreateUserAddrInput:
    properties:
      building_detail:
        type: string
      city_id:
        type: integer
      county_id:
        type: integer
      is_default:
        type: integer
      province_id:
        type: integer
      town_id:
        type: integer
    required:
    - building_detail
    - city_id
    - county_id
    - province_id
    - town_id
    type: object
  dto.DeleteCartEntriesInput:
    properties:
      cart_ids:
        items:
          type: integer
        type: array
    type: object
...
  dto.GetCartOutputElem:
    properties:
      avatar_url:
        description: 套餐头像url
        type: string
      hospital_id:
        description: 医院id
        type: integer
      hospital_name:
        description: 医院名字
        type: string
      id:
        description: 购物车条目id
        type: integer
      name:
        description: 套餐名字
        type: string
      pkg_count:
        description: 套餐数量
        type: integer
      pkg_id:
        description: 套餐id
        type: integer
      pkg_price:
        description: 套餐单价
        type: number
      update_time:
        description: 更新时间
        type: integer
    type: object
...
  dto.GetUserAddrOutput:
    properties:
      building_detail:
        type: string
      city_id:
        type: integer
      county_id:
        type: integer
      id:
        type: integer
      is_default:
        type: integer
      province_id:
        type: integer
      town_id:
        type: integer
    required:
    - building_detail
    - city_id
    - county_id
    - province_id
    - town_id
    type: object
  dto.JsApiTicketOutPut:
    properties:
      signature:
        description: 微信JsSDK签名
        type: string
    type: object
...
  dto.PostCartInput:
    properties:
      pkg_id:
        description: 要添加的套餐id
        type: integer
    type: object
...
  dto.ResourceID:
    properties:
      id:
        type: integer
    type: object
  dto.UpdateUserAddrInput:
    $ref: '#/definitions/dto.CreateUserAddrInput'
  dto.UserDetailOutput:
    properties:
      avatar_url:
        type: string
      gender:
        type: integer
      id:
        type: integer
      mobile:
        type: string
      open_id:
        type: string
      user_name:
        type: string
    type: object
  middleware.Response:
    properties:
      data:
        type: object
      ecode:
        type: integer
      emsg:
        type: string
    type: object
  model.UserAddr:
    properties:
      building_detail:
        type: string
      city_id:
        type: integer
      city_name:
        type: string
      county_id:
        type: integer
      county_name:
        type: string
      id:
        type: integer
      is_default:
        type: integer
      province_id:
        type: integer
      province_name:
        type: string
      town_id:
        type: integer
      town_name:
        type: string
      user_id:
        type: integer
    required:
    - building_detail
    - city_id
    - county_id
    - province_id
    - town_id
    type: object
info:
  contact: {}
  license: {}
paths:
  /cart/:
    delete:
      consumes:
      - application/json
      description: 删除购物车条目
      parameters:
      - description: 用户token
        in: header
        name: token
        required: true
        type: string
      - description: 加购的套餐id
        in: body
        name: cart_ids
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteCartEntriesInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 删除购物车条目
      tags:
      - cart
    get:
      consumes:
      - application/json
      description: 获取购物车里面的pkg列表
      parameters:
      - description: 用户token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.GetCartOutputElem'
                  type: array
              type: object
      summary: 获取购物车里面的pkg列表
      tags:
      - cart
    post:
      consumes:
      - application/json
      description: 加购物车
      parameters:
      - description: 用户token
        in: header
        name: token
        required: true
        type: string
      - description: 加购的套餐id
        in: body
        name: pkgId
        required: true
        schema:
          $ref: '#/definitions/dto.PostCartInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 往购物车增添一个套餐
      tags:
      - cart
...
  /users/addrs:
    get:
      description: 获取收件地址列表
      parameters:
      - description: 用户token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.UserAddr'
                  type: array
              type: object
      summary: 个人中心->账户信息->收件地址
      tags:
      - addrs
    post:
      consumes:
      - application/json
      description: 新增地址
      parameters:
      - description: 用户token
        in: header
        name: token
        required: true
        type: string
      - description: 新增用户收件地址
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserAddrInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ResourceID'
              type: object
      summary: 新增用户收件地址
      tags:
      - addrs
  /users/addrs/{id}:
    delete:
      consumes:
      - application/json
      description: 删除单个收件地址的
      parameters:
      - description: 用户token
        in: header
        name: token
        required: true
        type: string
      - description: addr id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ResourceID'
              type: object
      summary: 删除单个收件地址的
      tags:
      - addrs
    get:
      consumes:
      - application/json
      description: 获取单个收件地址的详情
      parameters:
      - description: 用户token
        in: header
        name: token
        required: true
        type: string
      - description: addr id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetUserAddrOutput'
              type: object
      summary: 获取单个收件地址的详情
      tags:
      - addrs
    put:
      consumes:
      - application/json
      description: 修改单个收件地址的
      parameters:
      - description: 用户token
        in: header
        name: token
        required: true
        type: string
      - description: addr id
        in: path
        name: id
        required: true
        type: integer
      - description: 修改用户收件地址
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserAddrInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ResourceID'
              type: object
      summary: 修改单个收件地址的
      tags:
      - addrs
  /users/profile:
    get:
      consumes:
      - application/json
      description: 获取用户的profile
      parameters:
      - description: 用户token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserDetailOutput'
              type: object
      summary: 个人中心->账户信息
      tags:
      - users
  /wx/js_ticket:
    get:
      consumes:
      - application/json
      description: 获取jsApiTicket 签名
      parameters:
      - description: 传入需要的调用js-sdk的uri地址
        in: query
        name: uri
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.JsApiTicketOutPut'
      summary: 获取签名
      tags:
      - WechatTag
...
swagger: "2.0"